requirements.txt
python-dotenv
telethon
pandas
sqlalchemy
psycopg2-binary
dbt-postgres
fastapi
uvicorn
ultralytics
dagster
dagster-webserver
opencv-python